pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

[project]
name = "cbcbeat"
version = "2019.1.2"
description = "An adjoint-enabled framework for computational cardiac electrophysiology"
authors = [{name = "M. E. Rognes", email = "meg@simula.no"}, {name = "J.E hake"}, {name = "P.E Farrel"}, {name = "S. W. Funke"}]
license = {file = "COPYING.LESSSER"}
readme = "README.md"
dependencies = [
    "fenics-dolfin",
    "numpy",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "matplotlib",
]
goss = [
    "pygoss",
]
gotran = [
    "gotran",
]
pypi = [
    "twine",
    "build"
]
docs = [
    "sphinx",
    "sphinx_book_theme",
]
all = [
   "cbcbeat[test]",
   "cbcbeat[docs]",
   "cbcbeat[pypi]",
   "cbcbeat[dev]"
]

# [project.scripts]
# drug-db = "scripts.gotran2"

[tool.setuptools]
license-files = ["COPYING"]
packages = ["cbcbeat", "cbcbeat.cellmodels"]
package-dir = {cbcbeat = "cbcbeat"}


[tool.pytest.ini_options]
addopts = [
   "--cov=cbcbeat",
   "--cov-report=html",
   "--cov-report=term-missing",
   "-v"
]
markers = [
    "fast: mark a test as fast",
    "medium: mark a test as medium ",
    "slow: mark a test as slow ",
    "adjoint: mark a test as using adjoint",
    "disabled: mark test as disabled",
    "xfail: mark test as failing ",
    "goss: mark test as using goss",
]
testpaths = [
    "test"
]

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = ["E402", "E741"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
